infrastructure/aws-cdk/requirements-test.txt
pytest>=7.4.0
pytest-cov>=4.1.0
aws-cdk-lib>=2.120.0
constructs>=10.3.0
